AutoWeb for Text inserted, Links are highlighted in Text and clickable. Delete and Forward repaired, ViewID and SQLite ID are different things. Action_View changed, a setData was added, so the Phone finds the right apps to show images or videos.

+++ app/src/main/java/de/radiohacks/frinmean/ChatFragment.java
@@ -92,7 +92,7 @@ public class ChatFragment extends ListFragment implements LoaderManager.LoaderCa
             setHasOptionsMenu(false);
             // Needed to show not the Chat where the Message is send from
             sendChatID = extras.getInt(Constants.SENDCHATID, -1);
-            MessageID = extras.getLong(Constants.SENDMSGID, -1);
+            MessageID = extras.getInt(Constants.SENDMSGID, -1);
         }
 
         SharedPreferences sharedPrefs = PreferenceManager.getDefaultSharedPreferences(getActivity());
